requirements.txt
amqp==2.6.1
asgiref==3.2.10
attrs==20.2.0
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.7
cffi==1.14.3
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
constantly==15.1.0
cryptography==3.1.1
cssselect==1.1.0
Django==3.1.2
django-celery-results==1.2.1
djangorestframework==3.12.1
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
itemadapter==0.1.1
itemloaders==1.0.3
jmespath==0.10.0
kombu==4.6.11
lxml==4.5.2
Markdown==3.3
parsel==1.6.0
prompt-toolkit==3.0.7
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
queuelib==1.5.0
Scrapy==2.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
vine==1.3.0
w3lib==1.22.0
wcwidth==0.2.5
zipp==3.3.0
zope.interface==5.1.2
